requirements.txt
fastapi
uvicorn
pydantic>=2
sqlalchemy
alembic
python-dotenv
python-docx
PyMuPDF
pdfplumber
sentence-transformers
faiss-cpu
jinja2
weasyprint
google-generativeai
passlib
python-jose[cryptography]
pytest
httpx
structlog
pytesseract
pillow
numpy
